requirements.txt
azure-common==1.1.28
azure-communication-email==1.0.0
azure-core==1.30.1
azure-mgmt-core==1.4.0
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
Flask==3.0.2
Flask-Cors==4.0.1
gevent==24.2.1
greenlet==3.0.3
idna==3.6
isodate==0.6.1
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
msrest==0.7.1
oauthlib==3.2.2
python-dotenv==1.0.1
requests==2.31.0
requests-oauthlib==2.0.0
setuptools==70.1.0
six==1.16.0
typing_extensions==4.11.0
urllib3==2.2.1
Werkzeug==3.0.2
zope.event==5.0
zope.interface==6.4.post2
